[Feature/#68] 그룹 홈/생성 API 연결

+++ domain/src/main/java/com/sopt/domain/entity/GroupEntity.kt
@@ -3,6 +3,6 @@ package com.sopt.domain.entity
 data class GroupEntity(
     val groupId: Long = -1,
     val groupName: String = "",
-    val groupMemberCount: Long = 0,
+    val groupMemberCount: Int = 0,
     val groupProfileImageUrl: String? = null
 )
